requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
langchain==0.1.9
langgraph==0.0.15
chromadb==0.4.22
nomic==3.0.12
pydantic==1.10.8
python-multipart==0.0.9
pdfminer.six==20231228
pytesseract==0.3.10
langsmith==0.1.22
ollama==0.1.6
sentence-transformers==2.5.1 